Big refactor - part 2 (end)
