+++ AuthorizeSetup/AddPermissionsToUserClaims.cs
@@ -10,7 +10,7 @@ using FeatureAuthorize;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
 
-namespace ServiceLayer.AuthorizeSetup
+namespace AuthorizeSetup
 {
     /// <summary>
     /// This version provides:
