+++ AuthorizeSetup/AuthCookieValidatePermissionsDataKey.cs
@@ -12,7 +12,7 @@ using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ServiceLayer.AuthorizeSetup
+namespace AuthorizeSetup
 {
     /// <summary>
     /// This version provides:
