+++ DataLayer/EfCode/ExtraAuthorizeDbContext.cs
@@ -4,12 +4,12 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using CommonCache;
 using DataLayer.EfCode.Configurations;
 using DataLayer.ExtraAuthClasses;
 using DataLayer.ExtraAuthClasses.Support;
 using DataLayer.MultiTenantClasses;
 using Microsoft.EntityFrameworkCore;
+using RefreshClaimsParts;
 
 namespace DataLayer.EfCode
 {
