+++ PermissionAccessControl2/Views/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using ServiceLayer.UserImpersonation
+@using UserImpersonation
 @inject SignInManager<IdentityUser> SignInManager
 @inject UserManager<IdentityUser> UserManager
