+++ Test/FakesAndMocks/FakeAuthChanges.cs
@@ -2,7 +2,7 @@
 // Licensed under MIT license. See License.txt in the project root for license information.
 
 using System;
-using CommonCache;
+using RefreshClaimsParts;
 
 namespace Test.FakesAndMocks
 {
