+++ Test/Test.csproj
@@ -47,11 +47,14 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DataAuthorize\DataAuthorize.csproj" />
+    <ProjectReference Include="..\DataKeyParts\DataKeyParts.csproj" />
     <ProjectReference Include="..\DataLayer\DataLayer.csproj" />
     <ProjectReference Include="..\FeatureAuthorize\FeatureAuthorize.csproj" />
     <ProjectReference Include="..\PermissionAccessControl2\PermissionAccessControl2.csproj" />
     <ProjectReference Include="..\PermissionParts\PermissionParts.csproj" />
+    <ProjectReference Include="..\RefreshClaimsParts\RefreshClaimsParts.csproj" />
     <ProjectReference Include="..\ServiceLayer\ServiceLayer.csproj" />
+    <ProjectReference Include="..\UserImpersonation\UserImpersonation.csproj" />
   </ItemGroup>
 
 </Project>
