+++ Test/UnitTests/FeatureAuthorizeTests/TestAuthChanges.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using CommonCache;
+using RefreshClaimsParts;
 using Test.FakesAndMocks;
 using Xunit;
 using Xunit.Extensions.AssertExtensions;
