+++ Test/UnitTests/FeatureAuthorizeTests/TestCacheRoleService.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Security.Claims;
-using CommonCache;
 using DataLayer.EfCode;
 using DataLayer.ExtraAuthClasses;
 using FeatureAuthorize;
