+++ Test/UnitTests/FeatureAuthorizeTests/TestPermissionsChangedFakeDatabase.cs
@@ -4,7 +4,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using CommonCache;
 using DataLayer.EfCode;
 using DataLayer.ExtraAuthClasses;
 using PermissionParts;
