+++ UserImpersonation/Concrete/ImpersonationCookie.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Http;
 
 [assembly: InternalsVisibleTo("Test")]
 
-namespace ServiceLayer.UserImpersonation.Concrete.Internal
+namespace UserImpersonation.Concrete
 {
     public class ImpersonationCookie
     {
