+++ UserImpersonation/Concrete/ImpersonationService.cs
@@ -1,13 +1,11 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Http;
-using ServiceLayer.UserImpersonation.Concrete.Internal;
 
-namespace ServiceLayer.UserImpersonation.Concrete
+namespace UserImpersonation.Concrete
 {
     public class ImpersonationService : IImpersonationService
     {
