+++ UserImpersonation/IImpersonationService.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
-namespace ServiceLayer.UserImpersonation
+namespace UserImpersonation
 {
     public interface IImpersonationService
     {
