+++ UserImpersonation/UserImpersonation.csproj
@@ -8,6 +8,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.DataProtection" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Html.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
+    <PackageReference Include="NetCore.AutoRegisterDi" Version="1.1.0" />
   </ItemGroup>
 
   <ItemGroup>
